Moved operations to background task; split AG calculation per competition
