+++ Competitie/models.py
@@ -76,12 +76,18 @@ DEELNAME_CHOICES = [(DEELNAME_ONBEKEND, 'Onbekend'),
                     (DEELNAME_JA, 'Bevestigd'),
                     (DEELNAME_NEE, 'Afgemeld')]
 
+MUTATIE_COMPETITIE_OPSTARTEN = 1
+MUTATIE_AG_VASTSTELLEN_18M = 2
+MUTATIE_AG_VASTSTELLEN_25M = 3
 MUTATIE_CUT = 10
 MUTATIE_INITIEEL = 20
 MUTATIE_AFMELDEN = 30
 MUTATIE_AANMELDEN = 40
 
-mutatie2descr = {
+MUTATIE_TO_STR = {
+    MUTATIE_AG_VASTSTELLEN_18M: "AG vaststellen 18m",
+    MUTATIE_AG_VASTSTELLEN_25M: "AG vaststellen 25m",
+    MUTATIE_COMPETITIE_OPSTARTEN: "competitie opstarten",
     MUTATIE_INITIEEL: "initieel",
     MUTATIE_CUT: "limiet aanpassen",
     MUTATIE_AFMELDEN: "afmelden",
@@ -711,7 +717,7 @@ class KampioenschapTeam(models.Model):
     klasse = models.ForeignKey(CompetitieKlasse, on_delete=models.CASCADE)
 
 
-class KampioenschapMutatie(models.Model):
+class KampioenschapMutatie(models.Model):       # TODO: hernoem naar CompetitieMutaties
 
     """ Deze tabel houdt de mutaties bij de lijst van (reserve-)schutters van
         de RK en BK wedstrijden.
@@ -761,11 +767,11 @@ class KampioenschapMutatie(models.Model):
         if not self.is_verwerkt:
             msg += " (nog niet verwerkt)"
         try:
-            msg += " %s (%s)" % (self.mutatie, mutatie2descr[self.mutatie])
+            msg += " %s (%s)" % (self.mutatie, MUTATIE_TO_STR[self.mutatie])
         except KeyError:
             msg += " %s (???)" % self.mutatie
 
-        if self.mutatie not in (MUTATIE_INITIEEL, MUTATIE_CUT):
+        if self.mutatie in (MUTATIE_AANMELDEN, MUTATIE_AFMELDEN):
             msg += " - %s" % self.deelnemer
 
         if self.mutatie == MUTATIE_CUT:
@@ -789,251 +795,7 @@ class CompetitieTaken(models.Model):
                                         on_delete=models.SET_NULL)
 
 
-def maak_deelcompetitie_ronde(deelcomp, cluster=None):
-    """ Maak een nieuwe deelcompetitie ronde object aan
-        geef er een uniek week nummer aan.
-    """
-
-    # zoek de bestaande records
-    objs = (DeelcompetitieRonde
-            .objects
-            .filter(deelcompetitie=deelcomp, cluster=cluster)
-            .order_by('-week_nr'))
-
-    # filter de import rondes eruit
-    objs = [obj for obj in objs if not obj.is_voor_import_oude_programma()]
-
-    if len(objs) > 0:
-        nieuwe_week_nr = objs[0].week_nr + 1
-
-        # maximum bereikt?
-        if len(objs) >= 10:
-            return
-    else:
-        nieuwe_week_nr = 37
-
-    # maak een eigen wedstrijdenplan aan voor deze ronde
-    plan = CompetitieWedstrijdenPlan()
-    plan.save()
-
-    ronde = DeelcompetitieRonde()
-    ronde.deelcompetitie = deelcomp
-    ronde.cluster = cluster
-    ronde.week_nr = nieuwe_week_nr
-    ronde.plan = plan
-    ronde.save()
-
-    return ronde
-
-
-def _maak_deelcompetities(comp, rayons, regios, functies):
-
-    """ Maak de deelcompetities van een competitie aan """
-
-    # zoek de voorgaande deelcompetities erbij om settings over te kunnen nemen
-    vorige_deelcomps = dict()   # [regio_nr] = DeelCompetitie()
-    for deelcomp in (DeelCompetitie
-                     .objects
-                     .select_related('competitie',
-                                     'nhb_regio')
-                     .filter(laag=LAAG_REGIO,
-                             competitie__begin_jaar=comp.begin_jaar - 1,
-                             competitie__afstand=comp.afstand)):
-        vorige_deelcomps[deelcomp.nhb_regio.regio_nr] = deelcomp
-    # for
-
-    # maak de Deelcompetities aan voor Regio, RK, BK
-    bulk = list()
-    for laag, _ in DeelCompetitie.LAAG:
-        if laag == LAAG_REGIO:
-            # Regio
-            for obj in regios:
-                functie = functies[("RCL", comp.afstand, obj.regio_nr)]
-                deel = DeelCompetitie(competitie=comp,
-                                      laag=laag,
-                                      nhb_regio=obj,
-                                      functie=functie)
-                try:
-                    vorige = vorige_deelcomps[obj.regio_nr]
-                except KeyError:
-                    pass
-                else:
-                    deel.inschrijf_methode = vorige.inschrijf_methode
-                    deel.toegestane_dagdelen = vorige.toegestane_dagdelen
-                    deel.regio_organiseert_teamcompetitie = vorige.regio_organiseert_teamcompetitie
-                    deel.regio_heeft_vaste_teams = vorige.regio_heeft_vaste_teams
-                    deel.regio_team_punten_model = vorige.regio_team_punten_model
-
-                bulk.append(deel)
-            # for
-        elif laag == LAAG_RK:
-            # RK
-            for obj in rayons:
-                functie = functies[("RKO", comp.afstand, obj.rayon_nr)]
-                deel = DeelCompetitie(competitie=comp,
-                                      laag=laag,
-                                      nhb_rayon=obj,
-                                      functie=functie)
-                bulk.append(deel)
-            # for
-        else:
-            # BK
-            functie = functies[("BKO", comp.afstand, 0)]
-            deel = DeelCompetitie(competitie=comp,
-                                  laag=laag,
-                                  functie=functie)
-            bulk.append(deel)
-    # for
-
-    DeelCompetitie.objects.bulk_create(bulk)
-
-
-def _maak_competitieklassen(comp, klassen_indiv, klassen_team):
-    """ Maak de competitieklassen aan voor een nieuwe competitie
-        het min_ag per klasse wordt later ingevuld
-    """
-
-    bulk = list()
-
-    for indiv in klassen_indiv:
-        klasse = CompetitieKlasse(
-                        competitie=comp,
-                        indiv=indiv,
-                        min_ag=AG_NUL)
-        bulk.append(klasse)
-    # for
-
-    for team in klassen_team:
-        klasse = CompetitieKlasse(
-                        competitie=comp,
-                        team=team,
-                        min_ag=AG_NUL)
-        bulk.append(klasse)
-    # for
-
-    CompetitieKlasse.objects.bulk_create(bulk)
-
-
-def competitie_aanmaken(jaar):
-    """ Deze functie wordt aangeroepen als de BKO de nieuwe competitie op wil starten
-        We maken de 18m en 25m competitie aan en daaronder de deelcompetities voor regio, rayon en bond
-
-        Wedstrijdklassen worden ook aangemaakt, maar het minimale AG wordt nog niet ingevuld
-    """
-    yearend = date(year=jaar, month=12, day=31)     # 31 december
-    udvl = date(year=jaar, month=8, day=1)          # 1 augustus
-    begin_rk = date(year=jaar + 1, month=2, day=1)  # 1 februari
-    begin_bk = date(year=jaar + 1, month=5, day=1)  # 1 mei
-
-    rayons = NhbRayon.objects.all()
-    regios = NhbRegio.objects.filter(is_administratief=False)
-    klassen_indiv = IndivWedstrijdklasse.objects.exclude(buiten_gebruik=True)
-    klassen_team = TeamWedstrijdklasse.objects.exclude(buiten_gebruik=True)
-
-    functies = dict()   # [rol, afstand, 0/rayon_nr/regio_nr] = functie
-    for functie in (Functie
-                    .objects
-                    .select_related('nhb_regio', 'nhb_rayon')
-                    .filter(rol__in=('RCL', 'RKO', 'BKO'))):
-        afstand = functie.comp_type
-        if functie.rol == 'RCL':
-            nr = functie.nhb_regio.regio_nr
-        elif functie.rol == 'RKO':
-            nr = functie.nhb_rayon.rayon_nr
-        else:  # elif functie.rol == 'BKO':
-            nr = 0
-
-        functies[(functie.rol, afstand, nr)] = functie
-    # for
-
-    # maak de Competitie aan voor 18m en 25m
-    for afstand, beschrijving in AFSTANDEN:
-        comp = Competitie(
-                    beschrijving='%s competitie %s/%s' % (beschrijving, jaar, jaar+1),
-                    afstand=afstand,      # 18/25
-                    begin_jaar=jaar,
-                    uiterste_datum_lid=udvl,
-                    begin_aanmeldingen=yearend,
-                    einde_aanmeldingen=yearend,
-                    einde_teamvorming=yearend,
-                    eerste_wedstrijd=yearend,
-                    laatst_mogelijke_wedstrijd=begin_rk,
-                    rk_eerste_wedstrijd=begin_rk,
-                    rk_laatste_wedstrijd=begin_rk + datetime.timedelta(days=7),
-                    bk_eerste_wedstrijd=begin_bk,
-                    bk_laatste_wedstrijd=begin_bk + datetime.timedelta(days=7))
-
-        if afstand == '18':
-            comp.laatst_mogelijke_wedstrijd = yearend
-
-        comp.save()
-
-        _maak_deelcompetities(comp, rayons, regios, functies)
-
-        _maak_competitieklassen(comp, klassen_indiv, klassen_team)
-    # for
-
-
-class KlasseBepaler(object):
-    """ deze klasse helpt met het bepalen van de CompetitieKlasse voor een deelnemer """
-
-    def __init__(self, competitie):
-        self.competitie = competitie
-        self.boogtype2klassen = dict()      # [boogtype.afkorting] = [klasse, klasse, ..]
-        self.lkl_cache_mannen = dict()      # [klasse.pk] = [lkl, lkl, ...]
-        self.lkl_cache_vrouwen = dict()     # [klasse.pk] = [lkl, lkl, ...]
-        self.lkl_cache = {'M': self.lkl_cache_mannen,
-                          'V': self.lkl_cache_vrouwen}
-
-        # vul de caches
-        for klasse in (CompetitieKlasse
-                       .objects
-                       .select_related('indiv',
-                                       'indiv__boogtype')
-                       .exclude(indiv=None)
-                       .prefetch_related('indiv__leeftijdsklassen')
-                       .all()):
-            indiv = klasse.indiv
-            boogtype = indiv.boogtype
-            try:
-                self.boogtype2klassen[boogtype.afkorting].append(klasse)
-            except KeyError:
-                self.boogtype2klassen[boogtype.afkorting] = [klasse]
-
-            for lkl in indiv.leeftijdsklassen.all():
-                lkl_cache = self.lkl_cache[lkl.geslacht]
-                try:
-                    lkl_cache[klasse.pk].append(lkl)
-                except KeyError:
-                    lkl_cache[klasse.pk] = [lkl]
-            # for
-        # for
-
-    def bepaal_klasse(self, aanmelding):
-        """ deze functie zet aanmelding.klasse aan de hand van de schutterboog """
-        ag = aanmelding.ag_voor_indiv
-        schutterboog = aanmelding.schutterboog
-        nhblid = schutterboog.nhblid
-        age = nhblid.bereken_wedstrijdleeftijd(self.competitie.begin_jaar + 1)
-
-        mogelijkheden = list()
-        for klasse in self.boogtype2klassen[schutterboog.boogtype.afkorting]:
-            if ag >= klasse.min_ag or klasse.indiv.is_onbekend:
-                for lkl in self.lkl_cache[nhblid.geslacht][klasse.pk]:
-                    if lkl.leeftijd_is_compatible(age):
-                        tup = (lkl, klasse)
-                        mogelijkheden.append(tup)
-                # for
-        # for
-
-        # indiv.volgorde heeft senioren eerst gezet; aspiranten laatste
-        # aspiranten mogen in alle klassen meedoen, maar we kiezen de Asp klasse
-        if len(mogelijkheden):
-            lkl, klasse = mogelijkheden[-1]
-            aanmelding.klasse = klasse
-        else:
-            # niet vast kunnen stellen
-            aanmelding.klasse = None
+# TODO: opruimen oude KampioenschapMutaties
 
 
 # end of file
