+++ Competitie/templates/competitie/kies.dtl
@@ -27,28 +27,16 @@
         {% include 'plein/card.dtl' with title="Uitslag vorig seizoen" text="De uitslag van vorig seizoen is hier in te zien." icon="skip_previous" url=card_url %}
     </div>
 
-    {% if bb_kan_competitie_aanmaken or bb_kan_ag_vaststellen %}
+    {% if bb_kan_competitie_aanmaken %}
         <h5>Management bondscompetitie</h5>
 
         <p>Deze mogelijkheid zijn alleen door de Manager Competitiezaken te gebruiken.</p>
 
         <div class="row">
-
             <!-- competities aanmaken -->
             {% if bb_kan_competitie_aanmaken %}
                 {% url 'Competitie:instellingen-volgende-competitie' as card_url %}
                 {% include 'plein/card.dtl' with title="Start competitie" text="Instellingen voor seizoen "|add:nieuwe_seizoen|add:" controleren en nieuwe competities aanmaken." icon="playlist_add" url=card_url %}
-            {% endif %}
-
-            <!-- aanvangsgemiddelden vaststellen -->
-            {% if bb_kan_ag_vaststellen %}
-                {% url 'Competitie:ag-vaststellen' as card_url %}
-                {% if bb_ag_nieuwste_datum %}
-                    <!-- template tag 'add' cannot handle missing variables -->
-                    {% include 'plein/card.dtl' with title="Aanvangsgemiddelden" text="AG vaststellen aan de hand van de uitslagen van vorig seizoen (op "|add:bb_ag_nieuwste_datum|add:" voor het laatst gedaan)" icon="how_to_reg" url=card_url %}
-                {% else %}
-                    {% include 'plein/card.dtl' with title="Aanvangsgemiddelden" text="AG vaststellen aan de hand van de uitslagen van vorig seizoen (eerste keer)" icon="how_to_reg" url=card_url %}
-                {% endif %}
             {% endif %}
         </div>
