+++ Competitie/views_overzicht.py
@@ -8,11 +8,11 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.http import Http404
 from django.views.generic import View, TemplateView
-from django.utils import timezone
 from django.utils.formats import localize
 from django.templatetags.static import static
+from Competitie.operations import bepaal_startjaar_nieuwe_competitie
 from Functie.rol import Rollen, rol_get_huidige, rol_get_huidige_functie, rol_get_beschrijving
-from Score.models import wanneer_ag_vastgesteld
+from Score.operations import wanneer_ag_vastgesteld
 from Taken.taken import eval_open_taken
 from .menu import menu_dynamics_competitie
 from .models import LAAG_REGIO, LAAG_BK, Competitie, DeelCompetitie
@@ -26,11 +26,6 @@ TEMPLATE_COMPETITIE_OVERZICHT_BEHEERDER = 'competitie/overzicht-beheerder.dtl'
 TEMPLATE_COMPETITIE_AANGEMELD_REGIO = 'competitie/lijst-aangemeld-regio.dtl'
 
 JA_NEE = {False: 'Nee', True: 'Ja'}
-
-
-def models_bepaal_startjaar_nieuwe_competitie():
-    """ bepaal het start jaar van de nieuwe competitie """
-    return timezone.now().year
 
 
 class CompetitieOverzichtView(View):
@@ -76,6 +71,15 @@ class CompetitieOverzichtView(View):
         if self.rol_nu == Rollen.ROL_BB:
             context['rol_is_bb'] = True
             kan_beheren = True
+
+            if not comp.klassegrenzen_vastgesteld:
+                afstand_meter = int(comp.afstand)
+                datum = wanneer_ag_vastgesteld(afstand_meter)
+                if datum:
+                    context['datum_ag_vastgesteld'] = localize(datum.date())
+                context['comp_afstand'] = comp.afstand
+                comp.url_ag_vaststellen = reverse('Competitie:ag-vaststellen-afstand',
+                                                  kwargs={'afstand': comp.afstand})
 
             context['planning_deelcomp'] = (DeelCompetitie
                                             .objects
@@ -362,12 +366,6 @@ class CompetitieKiesView(TemplateView):
                      .exclude(is_afgesloten=True)
                      .order_by('afstand', 'begin_jaar')):
 
-            if rol_nu == Rollen.ROL_BB:
-                if not comp.klassegrenzen_vastgesteld:
-                    # klassegrenzen zijn nog niet vastgesteld
-                    # laat de BB de aanvangsgemiddelden vaststellen
-                    context['bb_kan_ag_vaststellen'] = True
-
             comp.bepaal_openbaar(rol_nu)
 
             if comp.is_openbaar:
@@ -389,15 +387,9 @@ class CompetitieKiesView(TemplateView):
 
         if rol_nu == Rollen.ROL_BB:
             # als er nog geen competitie is voor het huidige jaar, geeft de BB dan de optie om deze op te starten
-            beginjaar = models_bepaal_startjaar_nieuwe_competitie()
+            beginjaar = bepaal_startjaar_nieuwe_competitie()
             context['nieuwe_seizoen'] = "%s/%s" % (beginjaar, beginjaar+1)
             context['bb_kan_competitie_aanmaken'] = (0 == Competitie.objects.filter(begin_jaar=beginjaar).count())
-
-            if 'bb_kan_ag_vaststellen' in context:
-                # zoek uit wanneer dit voor het laatste gedaan is
-                datum = wanneer_ag_vastgesteld()
-                if datum:
-                    context['bb_ag_nieuwste_datum'] = localize(datum.date())
 
         if rol_nu in (Rollen.ROL_IT, Rollen.ROL_BB, Rollen.ROL_BKO, Rollen.ROL_RKO, Rollen.ROL_RCL, Rollen.ROL_HWL):
             context['toon_beheerders'] = True
