+++ Schutter/test_regiocompetitie.py
@@ -10,9 +10,11 @@ from NhbStructuur.models import NhbRegio, NhbVereniging, NhbLid
 from Competitie.models import (Competitie, DeelCompetitie, RegioCompetitieSchutterBoog,
                                DeelcompetitieRonde, INSCHRIJF_METHODE_1, INSCHRIJF_METHODE_3)
 from Competitie.test_fase import zet_competitie_fase
+from Competitie.test_competitie import maak_competities_en_zet_fase_b, competities_aanmaken
 from Functie.models import Functie
 from Overig.e2ehelpers import E2EHelpers
-from Score.models import Score, ScoreHist, SCORE_TYPE_INDIV_AG, score_indiv_ag_opslaan
+from Score.models import Score, ScoreHist, SCORE_TYPE_INDIV_AG
+from Score.operations import score_indiv_ag_opslaan
 from Wedstrijden.models import CompetitieWedstrijd
 from .models import SchutterBoog
 import datetime
@@ -109,32 +111,7 @@ class TestSchutterRegiocompetitie(E2EHelpers, TestCase):
         # for
 
     def _competitie_aanmaken(self):
-        url_overzicht = '/bondscompetities/'
-        url_aanmaken = '/bondscompetities/aanmaken/'
-        url_ag_vaststellen = '/bondscompetities/ag-vaststellen/'
-        url_klassegrenzen_vaststellen = '/bondscompetities/%s/klassegrenzen/vaststellen/'   # comp_pk
-
-        # competitie aanmaken
-        with self.assert_max_queries(20):
-            resp = self.client.post(url_aanmaken)
-        self.assert_is_redirect(resp, url_overzicht)
-
-        comp_18 = Competitie.objects.get(afstand='18')
-        comp_25 = Competitie.objects.get(afstand='25')
-
-        # aanvangsgemiddelden vaststellen
-        with self.assert_max_queries(5):
-            resp = self.client.post(url_ag_vaststellen)
-
-        # klassegrenzen vaststellen
-        self.client.post(url_klassegrenzen_vaststellen % comp_18.pk)
-        self.client.post(url_klassegrenzen_vaststellen % comp_25.pk)
-
-        comp_18 = Competitie.objects.get(pk=comp_18.pk)
-        zet_competitie_fase(comp_18, 'B')
-
-        comp_25 = Competitie.objects.get(pk=comp_25.pk)
-        zet_competitie_fase(comp_25, 'B')
+        maak_competities_en_zet_fase_b()
 
     def test_inschrijven(self):
         # log in as BB en maak de competitie aan
