+++ Schutter/test_voorkeuren.py
@@ -9,7 +9,7 @@ from BasisTypen.models import BoogType
 from NhbStructuur.models import NhbRegio, NhbVereniging, NhbLid
 from Overig.e2ehelpers import E2EHelpers
 from Functie.models import maak_functie
-from Score.models import score_indiv_ag_opslaan
+from Score.operations import score_indiv_ag_opslaan
 from .models import SchutterBoog, SchutterVoorkeuren
 import datetime
