+++ Vereniging/test_lijst_rk.py
@@ -8,8 +8,8 @@ from django.test import TestCase
 from BasisTypen.models import BoogType
 from Functie.models import maak_functie
 from NhbStructuur.models import NhbRegio, NhbVereniging, NhbLid
-from Competitie.models import (Competitie, DeelCompetitie, CompetitieKlasse, LAAG_RK,
-                               DeelcompetitieKlasseLimiet, KampioenschapSchutterBoog)
+from Competitie.models import DeelCompetitie, CompetitieKlasse, LAAG_RK, DeelcompetitieKlasseLimiet, KampioenschapSchutterBoog
+from Competitie.test_competitie import maak_competities_en_zet_fase_b
 from Schutter.models import SchutterBoog
 from Overig.e2ehelpers import E2EHelpers
 import datetime
@@ -149,28 +149,8 @@ class TestVerenigingLijstRK(E2EHelpers, TestCase):
         self.e2e_wisselnaarrol_bb()
         self.e2e_check_rol('BB')
 
-        url_kies = '/bondscompetities/'
-        url_aanmaken = '/bondscompetities/aanmaken/'
-        url_klassegrenzen = '/bondscompetities/%s/klassegrenzen/vaststellen/'   # comp_pk
-
         self.assertEqual(CompetitieKlasse.objects.count(), 0)
-
-        with self.assert_max_queries(20):
-            resp = self.client.get(url_aanmaken)
-
-        # competitie aanmaken
-        with self.assert_max_queries(20):
-            resp = self.client.post(url_aanmaken)
-        self.assert_is_redirect(resp, url_kies)
-
-        self.comp_18 = Competitie.objects.get(afstand=18)
-        self.comp_25 = Competitie.objects.get(afstand=25)
-
-        # klassegrenzen vaststellen
-        resp = self.client.post(url_klassegrenzen % self.comp_18.pk)
-        self.assert_is_redirect(resp, url_kies)
-        resp = self.client.post(url_klassegrenzen % self.comp_25.pk)
-        self.assert_is_redirect(resp, url_kies)
+        self.comp_18, self.comp_25 = maak_competities_en_zet_fase_b()
 
         self.klasse_r = CompetitieKlasse.objects.filter(competitie=self.comp_18,
                                                         indiv__boogtype__afkorting='R',
